+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `--attribution` option to the `create` command
 - Add `delete` command for deleting a tileset
 - Add `tilejson` command
+- Add user-agent header to API requests
 
 # 1.0.1.dev0 (2020-04-07)
 - Fixed http status code for tilesets sources delete so it will no longer error